@@ -32,21 +32,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular-devkit/build-angular": "^0.901.4",
-    "@angular/animations": "^9.1.4",
-    "@angular/cdk": "^9.2.2",
-    "@angular/cli": "^9.1.4",
-    "@angular/common": "^9.1.4",
-    "@angular/compiler": "^9.1.4",
-    "@angular/compiler-cli": "^9.1.4",
-    "@angular/core": "^9.1.4",
-    "@angular/forms": "^9.1.4",
-    "@angular/localize": "^9.1.4",
-    "@angular/material": "^9.2.2",
-    "@angular/platform-browser": "^9.1.4",
-    "@angular/platform-browser-dynamic": "^9.1.4",
-    "@angular/platform-server": "^9.1.4",
-    "@angular/router": "^9.1.4",
+    "@angular/animations": "^10.0.2",
+    "@angular/cdk": "^10.0.1",
+    "@angular/cli": "^10.0.1",
+    "@angular/common": "^10.0.2",
+    "@angular/compiler": "^10.0.2",
+    "@angular/core": "^10.0.2",
+    "@angular/forms": "^10.0.2",
+    "@angular/localize": "^10.0.2",
+    "@angular/material": "^10.0.1",
+    "@angular/platform-browser": "^10.0.2",
+    "@angular/platform-browser-dynamic": "^10.0.2",
+    "@angular/platform-server": "^10.0.2",
+    "@angular/router": "^10.0.2",
     "@ngrx/effects": "^9.1.1",
     "@ngrx/entity": "^9.1.1",
     "@ngrx/router-store": "^9.1.1",
@@ -54,15 +52,15 @@
     "@ngrx/store-devtools": "^9.0.0",
     "dotenv": "^8.2.0",
     "intersection-observer": "^0.10.0",
-    "rxjs": "^6.5.5",
-    "typescript": "^3.9.6",
-    "zone.js": "^0.10.3"
+    "rxjs": "^6.6.0",
+      "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-builders/custom-webpack": "^9.2.0",
-    "@angular-builders/dev-server": "^7.3.1",
-    "@angular/language-service": "^9.1.4",
+    "@angular-devkit/build-angular": "^0.1000.1",
+    "@angular/compiler-cli": "^10.0.2",
+    "@angular/language-service": "^10.0.2",
     "@ngrx/store-devtools": "^9.1.1",
+    "@ngtools/webpack": "^10.0.1",
     "@types/express": "^4.17.6",
     "@types/hammerjs": "^2.0.36",
     "@types/jasmine": "~3.5.10",
@@ -84,6 +82,9 @@
     "protractor": "~5.4.4",
     "ts-loader": "^7.0.2",
     "ts-node": "^8.10.1",
+    "tslib": "^2.0.0",
+    "tslint": "^6.1.2",
+    "typescript": "^3.9.6",
     "webpack": "^4.43.0",
     "webpack-cli": "^3.3.11"
   }
