@@ -1,4 +1,5 @@
 const webpack = require('webpack');
+const { AngularCompilerPlugin } = require('@ngtools/webpack');
 module.exports = {
   plugins: [
     new webpack.DefinePlugin({
@@ -6,6 +7,12 @@ module.exports = {
         'AWS_ACCESS_KEY_ID' : JSON.stringify(process.env.AWS_ACCESS_KEY_ID),
         'AWS_BUCKET_NAME' : JSON.stringify('smc-static-media'),
         'AWS_END_POINT' : JSON.stringify('http://smc-static-media.s3-website.eu-west-3.amazonaws.com')
-    })
+    }),
+    // new AngularCompilerPlugin({
+    //   "tsConfigPath": "./tsconfig.json",
+    //   compilerOptions: {
+    //     "enableIvy": true,
+    //   }
+    // })
   ]
 };
